// TaxiServiceSPA/src/interfaces/models/operator.ts
export interface Operator {
    operatorId: number;
    departmentId: number;
    lastName: string;
    firstName: string;
    middleName: string;
    dateOfBirth: string;
    region: string;
    city: string;
    street: string;
    building: string;
    flat: number;
    beginning: string;
    ending: string;
    salary: number;
    workingPhone: string;
}